Export a proper library from `staking-cli`
`staking-cli` has a lot of functions we can use in `TestNetwork`. In
order to use them we need it to act as a library. This PR also adds
the lib to `sequencer` crate for use in follow up PRs.
