+++ Cargo.toml
@@ -53,6 +53,7 @@ exclude = ["sequencer-sqlite"]
 [workspace.dependencies]
 contract-bindings-alloy = { path = "./contract-bindings-alloy" }
 ethers-conv = { path = "./ethers-conv" }
+staking-cli = { path = "./staking-cli" }
 alloy = { version = "0.9", default-features = false, features = [
     "contract",
     "eips",
