+++ sequencer/Cargo.toml
@@ -53,6 +53,7 @@ async-once-cell = { workspace = true }
 async-trait = { workspace = true }
 bincode = { workspace = true }
 byteorder = "1"
+staking-cli = { workspace = true }
 
 # CDN imports
 cdn-broker = { git = "https://github.com/EspressoSystems/Push-CDN", tag = "0.5.1-upgrade", package = "cdn-broker", features = ["global-permits"] }
